improve merge_duplicate_records.rb

+++ lib/tasks/merge_duplicate_records.rb
@@ -1,61 +1,85 @@
 # frozen_string_literal: true
 
-# To do:
-#   - deal with "The" at the beginning of titles -- this should happen in a fuzzy matcher
-#   - incorporate fuzzy matching to find possible duplicates
-#     - "booker t and the mg's" vs "booker t and the mgs"
-#   - Create a CSV of fuzzy matches to review?
-#   - look up preferred names from Spotify API
-
-# Incorporate fuzzy matching:
-# (4..9).map { |threshold| Artist.where("similarity(name, ?) > ?", "booker t and the mgs", (threshold / 10.0)).count }
-# For booker t, .4 is at the low end of usefulness, but does find mostly relevant results; .5 finds all relevant results
-
-# This helper class merges records that are invalid due to duplicate normalized columns
+# Merge duplicate records that are invalid due to duplicate normalized columns
 # The normalized column is then populated for the primary record
-class MergeDuplicateRecords
-  attr_reader :model, :column_name, :normalized_column_name
 
-  def initialize(model, column_name, normalized_column_name = nil)
-    @model = model
+# Classes that will be normalized:
+# Normalizable::NORMALIZED_CLASSES_AND_COLUMNS.keys.map do |class_name|
+#   children = class_name.singularize.camelize.constantize.reflections.select do |_key, reflection|
+#     reflection.class.in?(WHITELISTED_REFLECTION_CLASSES) && !reflection.options[:through]
+#   end
+#
+#   [class_name, children.values.map{ |v| [v.name, v.class.name.split("::").last] }.to_h]
+# end.to_h
+# =>
+# {"Album"=>{:albums_songs=>"HasManyReflection"},
+#  "Artist"=>{:albums=>"HasManyReflection", :songs=>"HasManyReflection"},
+#  "Broadcast"=>{:playlists=>"HasManyReflection"},
+#  "Genre"=>{:songs=>"HasManyReflection"},
+#  "Playlist"=>{:playlists_songs=>"HasManyReflection", :playlist_import=>"HasOneReflection"},
+#  "RecordLabel"=>{:albums=>"HasManyReflection"},
+#  "Song"=>{:playlists_songs=>"HasManyReflection", :albums_songs=>"HasManyReflection"}}
+
+class MergeDuplicateRecords
+  attr_reader :klass, :column_name, :normalized_column_name
+
+  WHITELISTED_REFLECTION_CLASSES = [
+    ActiveRecord::Reflection::HasManyReflection,
+    ActiveRecord::Reflection::HasOneReflection
+  ].freeze
+
+  def initialize(klass, column_name, normalized_column_name = nil)
+    @klass = klass
     @column_name = column_name
     # All normalized columns are named "normalized_#{column_name}", but can be overridden if necessary:
-    @normalized_column_name = normalized_column_name || "normalized_#{column_name}"
+    @normalized_column_name = normalized_column_name || "#{Normalizable::NORMALIZED_PREFIX}#{column_name}"
   end
 
   def perform
-    unless model.column_names.include?(normalized_column_name) && model.column_names.include?(column_name)
-      raise "Model #{model.name} does not have the necessary columns: #{column_name} and #{normalized_column_name}"
+    unless klass.column_names.include?(normalized_column_name) && klass.column_names.include?(column_name)
+      raise "Class #{klass.name} does not have the necessary columns: #{column_name} and #{normalized_column_name}"
     end
 
-    find_possible_duplicate_id_groups
-      .each { |hash| merge_records(hash) }
+    find_possible_duplicate_id_groups.each do |normalized_name, record_ids|
+      merge_records(normalized_name, record_ids)
+    end
   end
 
   def update_id_column(record, new_id)
     # Using update_column to bypass validations; some records are no longer valid due to the normalized columns
-    record.update_column("#{model.name.underscore}_id", new_id) # rubocop:disable Rails/SkipsModelValidations
+    record.update_column("#{klass.name.underscore}_id", new_id) # rubocop:disable Rails/SkipsModelValidations
   end
 
-  def update_children(parent_record, new_id) # rubocop:disable Metrics/MethodLength
-    model.reflections.each_key do |association_name|
-      associated_records = parent_record.send(association_name)
+  def update_children(parent_record, new_id)
+    klass.reflections.each_value do |reflection|
+      next unless reflection.class.in?(WHITELISTED_REFLECTION_CLASSES)
+      next if reflection.options[:through]
 
-      ActiveRecord::Base.transaction do
-        # Check if the association is a collection or a single record
-        if associated_records.respond_to?(:each)
-          associated_records.each do |record|
-            update_id_column(record, new_id)
-          end
-        else
-          update_id_column(associated_records, new_id)
-        end
+      Array.wrap(parent_record.send(reflection.name)).each do |child|
+        update_id_column(child, new_id)
+      end
+    end
+  end
+
+  # This is redundant, but I wanted to include a sanity check for peace of mind
+  def ensure_children_have_correct_id(parent_record, new_id)
+    klass.reflections.each_value do |reflection|
+      next unless reflection.class.in?(WHITELISTED_REFLECTION_CLASSES)
+      next if reflection.options[:through]
+
+      Array.wrap(parent_record.send(reflection.name)).each do |child|
+        next if child.send("#{klass.name.underscore}_id") == new_id
+
+        raise <<~ERROR_MSG
+          Child record #{child.class.name} #{child.id} has incorrect #{klass.name.underscore}_id: #{child.send("#{klass.name.underscore}_id")}
+          Parent record #{parent_record.class.name} #{parent_record.id} has #{klass.name.underscore}_id: #{new_id}
+        ERROR_MSG
       end
     end
   end
 
   def find_possible_duplicate_id_groups
-    groups = model.all.each_with_object(Hash.new { |h, k| h[k] = [] }) do |record, hash|
+    groups = klass.all.each_with_object(Hash.new { |h, k| h[k] = [] }) do |record, hash|
       existing_name = record.send(column_name)
       normalized_name = Normalizable.normalize_text(existing_name)
       hash[normalized_name] << record.id
@@ -64,32 +88,28 @@ class MergeDuplicateRecords
     groups.select { |_k, v| v.size > 1 }
   end
 
-  def choose_primary_record(records, _normalized_column_name)
-    puts 'Select the primary record for the following group:' # rubocop:disable Rails/Output
-    records.each_with_index { |record, index| puts "#{index + 1}. #{record.send(column_name)}" }
-
-    puts 'Primary record number: ' # rubocop:disable Rails/Output
-    primary_record_number = gets.chomp.to_i
-    records[primary_record_number - 1]
-
-    # Too tedious to do manually.
-    # TODO: Implement NameFormatter to automate selecting the 'best' name for most cases.
-    # NameFormatter.format_name(records.map(&column_name))
+  def choose_primary_record(records)
+    name_array = records.map { |r| r.send(column_name) }
+    best_name = NameFormatter.format_name(name_array)
+    records.select { |record| record.send(column_name) == best_name }.first
   end
 
-  def merge_records(hash)
-    normalized_column_name, id_group = hash
-    # prompt user to select the primary record
-    records = model.find(id_group)
-    primary_record = choose_primary_record(records, normalized_column_name)
+  def merge_records(normalized_name, record_ids)
+    records = klass.find(record_ids)
+    primary_record = choose_primary_record(records)
+    return unless primary_record
 
-    records.each do |record|
-      next if record.id == primary_record.id
+    ActiveRecord::Base.transaction(requires_new: true) do
+      records.each do |record|
+        next if record.id == primary_record.id
 
-      update_children(record, primary_record.id)
-      record.destroy!
+        update_children(record, primary_record.id)
+        ensure_children_have_correct_id(primary_record, primary_record.id)
+        record.destroy!
+      end
+      # populate the normalized column for the primary record
+      primary_record.send("#{normalized_column_name}=", normalized_name)
+      primary_record.save!
     end
-    # populate the normalized column for the primary record
-    primary_record.save!
   end
 end
